requirements.txt
Flask==3.0.3
requests==2.32.5
paho-mqtt==2.1.0